requirements.txt
requests==2.31.0
python-dotenv==1.0.0
markdown==3.5.1
reportlab==4.0.7
streamlit==1.28.1